Extracted project/site components to folder
