+++ src/app/components/projects/components/site-card/site-card.component.scss
