+++ src/app/components/projects/projects.module.ts
@@ -3,8 +3,11 @@ import { RouterModule } from "@angular/router";
 import { getRouteConfigForPage } from "@helpers/page/pageRouting";
 import { MapModule } from "@shared/map/map.module";
 import { SharedModule } from "@shared/shared.module";
-import { HarvestCompleteComponent } from "./harvest-complete/harvest-complete.component";
-import { HarvestReviewComponent } from "./harvest-review/harvest-review.component";
+import { HarvestCompleteComponent } from "./components/harvest-complete/harvest-complete.component";
+import { HarvestReviewComponent } from "./components/harvest-review/harvest-review.component";
+import { PillListComponent } from "./components/pill-list/pill-list.component";
+import { SiteCardComponent } from "./components/site-card/site-card.component";
+import { SiteMapComponent } from "./components/site-map/site-map.component";
 import { AssignComponent } from "./pages/assign/assign.component";
 import { DeleteComponent } from "./pages/delete/delete.component";
 import { DetailsComponent } from "./pages/details/details.component";
@@ -14,10 +17,7 @@ import { ListComponent } from "./pages/list/list.component";
 import { NewComponent } from "./pages/new/new.component";
 import { PermissionsComponent } from "./pages/permissions/permissions.component";
 import { RequestComponent } from "./pages/request/request.component";
-import { PillListComponent } from "./pill-list/pill-list.component";
 import { projectsRoute } from "./projects.menus";
-import { SiteCardComponent } from "./site-card/site-card.component";
-import { SiteMapComponent } from "./site-map/site-map.component";
 
 const components = [
   AssignComponent,
